Use ALT-UP and ALT-DOWN to go to parent/child directories in dir view
CAUTION: The way I implemented this is very stupid. Because ALT-DOWN and
ALT-UP are actually accelerators for next diff and prev diff both in dir
view and in editor view, I couldn't find a way to change ALT-arrow
shortcuts for only dir views. Therefore I had to break the next diff and
prev diff buttons' behavior. If we're in the dir view, those buttons now
are used for traversing directories.